feat(ui): rename LxdProject to LXDHostVMs and make reusable
This branch renames LxdProject to LXDHostVMs and makes it possible to be reused
across LXD cluster and LXD single host pages.
