+++ ui/src/app/kvm/components/LXDHostVMs/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LXDHostVMs";
