+++ ui/src/app/kvm/components/LXDVMsSummaryCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LXDVMsSummaryCard";
