+++ ui/src/app/kvm/components/LXDVMsTable/VMsActionBar/_index.scss
