+++ ui/src/app/kvm/components/LXDVMsTable/VMsActionBar/index.ts
