+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/CoresColumn/CoresColumn.tsx
