+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/CoresColumn/index.ts
