+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/HugepagesColumn/HugepagesColumn.test.tsx
