+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/HugepagesColumn/HugepagesColumn.tsx
