+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/HugepagesColumn/index.ts
