+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/IPColumn/IPColumn.test.tsx
