+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/NameColumn/NameColumn.test.tsx
