+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/NameColumn/index.ts
