+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/_index.scss
