+++ ui/src/app/kvm/components/LXDVMsTable/VMsTable/index.ts
