+++ ui/src/app/kvm/components/LXDVMsTable/index.ts
@@ -1,0 +1,1 @@
+export { default, VMS_PER_PAGE } from "./LXDVMsTable";
