--- ui/src/app/kvm/components/LxdProject/LxdProject.test.tsx
@@ -1,192 +1,0 @@
-import { mount } from "enzyme";
-import { Provider } from "react-redux";
-import { MemoryRouter } from "react-router-dom";
-import configureStore from "redux-mock-store";
-
-import LxdVMs from "../LxdVMs";
-
-import LxdProject from "./LxdProject";
-
-import kvmURLs from "app/kvm/urls";
-import { PodType } from "app/store/pod/constants";
-import {
-  machine as machineFactory,
-  pod as podFactory,
-  podPowerParameters as powerParametersFactory,
-  podResources as podResourcesFactory,
-  podState as podStateFactory,
-  podVM as podVMFactory,
-  rootState as rootStateFactory,
-} from "testing/factories";
-
-const mockStore = configureStore();
-
-describe("LxdProject", () => {
-  it("shows a spinner if pods have not loaded yet", () => {
-    const state = rootStateFactory({
-      pod: podStateFactory({
-        items: [],
-        loaded: false,
-      }),
-    });
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/project", key: "testKey" }]}
-        >
-          <LxdProject
-            id={1}
-            searchFilter=""
-            setSearchFilter={jest.fn()}
-            setHeaderContent={jest.fn()}
-          />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    expect(wrapper.find("Spinner").exists()).toBe(true);
-  });
-
-  it("shows a message if the pod can't be found", () => {
-    const state = rootStateFactory({
-      pod: podStateFactory({
-        items: [],
-        loaded: true,
-      }),
-    });
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/project", key: "testKey" }]}
-        >
-          <LxdProject
-            id={1}
-            searchFilter=""
-            setSearchFilter={jest.fn()}
-            setHeaderContent={jest.fn()}
-          />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    expect(wrapper.find("[data-test='not-found']").exists()).toBe(true);
-  });
-
-  it("redirects to Virsh pod details page if pod is not a LXD pod", () => {
-    const state = rootStateFactory({
-      pod: podStateFactory({
-        items: [podFactory({ id: 1, type: PodType.VIRSH })],
-        loaded: true,
-      }),
-    });
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[
-            { pathname: kvmURLs.lxd.single.vms({ id: 1 }), key: "testKey" },
-          ]}
-        >
-          <LxdProject
-            id={1}
-            searchFilter=""
-            setSearchFilter={jest.fn()}
-            setHeaderContent={jest.fn()}
-          />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    expect(wrapper.find("Redirect").exists()).toBe(true);
-    expect(wrapper.find("Redirect").prop("to")).toBe(
-      kvmURLs.virsh.details.index({ id: 1 })
-    );
-  });
-
-  it("can display the project name", () => {
-    const state = rootStateFactory({
-      pod: podStateFactory({
-        items: [
-          podFactory({
-            id: 1,
-            power_parameters: powerParametersFactory({
-              project: "blair-witch",
-            }),
-            type: PodType.LXD,
-          }),
-        ],
-        loaded: true,
-      }),
-    });
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/project", key: "testKey" }]}
-        >
-          <LxdProject
-            id={1}
-            searchFilter=""
-            setSearchFilter={jest.fn()}
-            setHeaderContent={jest.fn()}
-          />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    expect(wrapper.find("[data-test='project-name']").text()).toBe(
-      "blair-witch"
-    );
-  });
-
-  it("can get resources for a VM", () => {
-    const state = rootStateFactory({
-      pod: podStateFactory({
-        items: [
-          podFactory({
-            id: 1,
-            resources: podResourcesFactory({
-              vms: [
-                podVMFactory({
-                  hugepages_backed: true,
-                  pinned_cores: [3],
-                  system_id: "abc123",
-                  unpinned_cores: 8,
-                }),
-              ],
-            }),
-            type: PodType.LXD,
-          }),
-        ],
-        loaded: true,
-      }),
-    });
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/project", key: "testKey" }]}
-        >
-          <LxdProject
-            id={1}
-            searchFilter=""
-            setSearchFilter={jest.fn()}
-            setHeaderContent={jest.fn()}
-          />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    expect(
-      wrapper.find(LxdVMs).invoke("getResources")(
-        machineFactory({ system_id: "abc123" })
-      )
-    ).toStrictEqual({
-      hugepagesBacked: true,
-      pinnedCores: [3],
-      unpinnedCores: 8,
-    });
-  });
-});
