--- ui/src/app/kvm/components/LxdProject/ProjectSummaryCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ProjectSummaryCard";
