--- ui/src/app/kvm/components/LxdProject/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./LxdProject";
