--- ui/src/app/kvm/components/LxdVMs/index.ts
@@ -1,1 +1,0 @@
-export { default, VMS_PER_PAGE } from "./LxdVMs";
